+++ src/main/java/com/odorok/OdorokApplication/infrastructures/domain/Sido.java
@@ -1,28 +1,21 @@
-package com.odorok.OdorokApplication.draftDomain;
+package com.odorok.OdorokApplication.infrastructures.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "sigungus")
+@Entity(name = "sidos")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
-public class Sigungu {
+public class Sido {
     @Id
-    private Integer no;
+    @Column(name = "code")
+    private Integer code;
 
     @Column(name= "name")
     private String name;
-
-    @Column(name= "sido_code")
-    private String sidoCode;
-
-    @Column(name= "code")
-    private String code;
 }
