+++ src/main/java/com/odorok/OdorokApplication/infrastructures/repository/SidoRepository.java
@@ -1,6 +1,6 @@
 package com.odorok.OdorokApplication.infrastructures.repository;
 
-import com.odorok.OdorokApplication.draftDomain.Sido;
+import com.odorok.OdorokApplication.infrastructures.domain.Sido;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
