+++ src/main/java/com/odorok/OdorokApplication/infrastructures/repository/SigunguRepository.java
@@ -1,6 +1,6 @@
 package com.odorok.OdorokApplication.infrastructures.repository;
 
-import com.odorok.OdorokApplication.draftDomain.Sigungu;
+import com.odorok.OdorokApplication.infrastructures.domain.Sigungu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
