+++ src/main/java/com/odorok/OdorokApplication/infrastructures/service/DurunubiDataServiceImpl.java
@@ -5,8 +5,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.odorok.OdorokApplication.commons.webflux.util.WebClientUtil;
-import com.odorok.OdorokApplication.draftDomain.Sido;
-import com.odorok.OdorokApplication.draftDomain.Sigungu;
+import com.odorok.OdorokApplication.infrastructures.domain.Sido;
+import com.odorok.OdorokApplication.infrastructures.domain.Sigungu;
 import com.odorok.OdorokApplication.infrastructures.domain.*;
 import com.odorok.OdorokApplication.infrastructures.repository.CourseRepository;
 import com.odorok.OdorokApplication.infrastructures.repository.PathCoordRepository;
