+++ src/main/java/com/odorok/OdorokApplication/infrastructures/service/RegionServiceTempImpl.java
@@ -1,7 +1,7 @@
 package com.odorok.OdorokApplication.infrastructures.service;
 
-import com.odorok.OdorokApplication.draftDomain.Sido;
-import com.odorok.OdorokApplication.draftDomain.Sigungu;
+import com.odorok.OdorokApplication.infrastructures.domain.Sido;
+import com.odorok.OdorokApplication.infrastructures.domain.Sigungu;
 import com.odorok.OdorokApplication.infrastructures.repository.SidoRepository;
 import com.odorok.OdorokApplication.infrastructures.repository.SigunguRepository;
 import lombok.RequiredArgsConstructor;
